+++ Emulator/Agnus/Blitter/FastBlitter.cpp
@@ -356,7 +356,7 @@ Blitter::doFastLineBlit()
     bltcon = bltcon & 0x0FFFFFFBF;
     if (decision_is_signed) bltcon |= 0x00000040;
     
-    setBLTCON0ASH(blit_a_shift_local);
+    setBLTCON0ASH((u16)blit_a_shift_local);
     bnew   = bltbdat_local;
     
     bltapt = decision_variable & agnus.ptrMask;
